Zigzag Harris energy reduction by a factor of 100
This point is not closed but part of the issue with what I was
doing before was using the midpoints and normalized displacement
calculated after the comment: // Compute the midpoint and the
normalized displacment of the streak. I used them as if they were
the initial point and total displacement of the particle.

This does not solve all the problems, but it represents a start.
Currently, accumulate_j_zigzag.hpp is full of "joe_" variables so
as not to interfere with s_midx, s_dispx, etc., in the rest of
move_p_kokkos(...).