In the process of debugging
And so printf statements are flying around like crazy. One thing
that I caught was the in the reference point calculation, I forgot
to add a constant 1.f inside the outermost fmin(...). This is
fixed and the reference point now appears to be correct. The
numerical values for the energies and fields are again within an
order of magnitude and usually within 10% of the standard move_p_
kokkos(...) implementation. I'm not yet sure where the final
discrepancy is from though.