Still debugging the current accumulation
Something is weird regarding the in-cell coordinates compared
to the Umeda paper. For some reason current is not being
accumulated on all parts of the cell where it should be -- right
now it's only accumulated on the v0 edge. Also, for some reason
move_p_kokkos(...) is being entered when the particle does not
leave the cell.

The main changes right now are print statements outputting the
current accumulation values, and I reverted the changes in the
particle position and displacements for simplicity (they changes
are currently commented out).