Cell and Umeda coordinates conversion macro-ed up
A quick hack resulting in some progress along the current
accumulation lines has been done by converting between cell
coordinates which run from [-1,1] and the coordinates assumed in
Umeda et al, "Umeda coordinates", [0,1]. The current using this
method along the zig of the zigzag is regularly between a factor
of 5-10 smaller than the older current accumulation method.
However the 2particle.cxx deck is almost identical through 500
time steps. This is may be due to the deck's simplicity.

Ignore giant swaths of commented-out code. That will be removed
once these macro hacks are Bob-certified and promoted to
officially-sanctioned code. (They are left in for the sake of
future discussion.)

It still needs to be verified that the "two zig" method for the
zigzag actually is a zigzag and not just wishful thinking.