Trying something new next
The coordinate conversions, included those for the fluxes, seem
weird. I don't like the random factors of 8. Also, currents were
not being deposited at all on the zag, so I added quite a few if
statements to turn on l = 2.

At this point I'm going to try something different and run away
from the dumpster fire this code's becoming. If I can't figure it
out with the new idea with the weights, then come back to this
commit.