Moved some things around and started zigzag again
Since it looks like the particle weight factors between the
Villasenor and Umeda algorithms are almost the same (up to that
extra 1/12 term in the Villasenor paper), the idea now is to move
the particles manually along the zigzag, and then have the
current accumulate over this new trajectory.

The end result is a lot of if statements that should be, in
principle, possible to cut down once we know the idea works. The
code runs and spits out the correct 1d zig currents, but fails on
the 1d zag currents. I haven't tried higher dimensions. I also
added more comments to the older code to help future readers
understand move_p_kokkos(...) without as much difficulty.

Two things that will need to be addressed. First, even in the old
code, it looks the the particle displacement is unnecessarily
multiplied by a factor of 2 when axis == 3. This I do not quite
get yet, and so I don't know how important it is. Second, it is
not clear how to shoot a particle through a corner instead of only
along one cube face as it is done now. If this cannot be easily
done, it's not clear that the zigzag algorithm will actually be
any different than what's currently implemented in the code.