Zig and Zag seem to be working
The zigzag method works when s_disp is taken to be half of the
actual displacement that the particle is going to move along as
calculated by advance_p. Additionally, the displacements should
only be further halved if the particle leaves the cell -- that
is, when axis != 3. The 2particle deck accumulates the same
currents now between the two methods now.

The particles now must be properly moved to the boundaries such
that edge or corner crossings are allowed.