Fixed the segfault
In init_kokkos_grid(...) in initialize.cc on devel, the size of
the neighbor array was set as the input argument. After changing
the size of the local neighbors from 6 to 27, this input argument
was taken out but it needed to be 27 * number_of_voxels. This is
"fixed" now, and so I can continue figuring out why the energies
are off.