Removed some comments and finished the functor
Now the neighbor indices can be plotted as histograms which show
the number of instances a particle is moved to a particular
neighbor with index 0, 1, 2, ..., 26.

Additionally, there is no discrepancy in the energies.txt files
for the harris deck between devel and the Villasenor method with
27 neighbors on this branch. Currently zigzag does not agree with
these methods. Interestingly enough, there are more neighbor
indices for zigzag than there are for Villasenor implying that
there are more move attempts in zigzag. This seems wrong...