Added final destination connection writer
This poor VPIC_DUMP_NEIGHBORS "functor" is really more of just an
output stream object at this point. Anyway it now determines if
each particle moves through a Face, Edge, or Corner and writes
that info to a file for future analysis. It's smart enough to only
do this once per particle to avoid over-counting. This can be used
to estimate what the likely speedup zigzag will have over the
straight-line method which can only send particles through faces.