Update on zigzag bug
The bug was not that the midpoint was wrong, it was that the
current accumulation was not calculated at the "quarter-point", as
expected in the accumulate_j macro. Instead what was happening is
the currents were accumulated relative to the full displacement
rather than the displacement to the midpoint first along the zig
before proceeding to the zag.