Displacements need to be inside IFs
These are the same IFs that I'm pretty sure I can eliminate by
merging them with ones that come before. So far it looks like the
particle positions are what they should be and the midpoints along
the zig and zag (so-called "quarter-points") are used to calculate
the current accumulations. So things look promising right now!

There are still systematic discrepancies that are expected for
large numbers of particles moving through edges and corners. A more
thorough analysis is needed to test the veracity/fidelity of VPIC
with "working" zigzags.