Harris agrees very wel for a single time step
After the first time step, the particle energies in the Harris
deck seem to be different only in the 5 significant figure. This
is heartening, as it means that the problems again may be due to
the lack of periodic boundary conditions with extended neighbors.

For another test, I will now change advance_p so eliminate any
contribution from the fields to study the particle motion.