Cleaned up zigzag code a bit for readability
It's still hard to tell the degree of wrongness associated with
zigzags. It looks like the particles are moving along the proper
paths and depositing current appropriately. There is still a global
periodic boundary condition problem thanks to GLUE_FACE and SET_PBC,
but I'm not so sure how relevant they are.