Properly tracks rhob between host and device.
The host now has an accumulator array for rhob that tracks changes from
the absorbing particle boundary condition.  This is copied to the device
which then adjusts its rhob appropriately.

There is still a copy of an unnecessarily large array, so doing all of
the absorbing boundary conditions on the device is preferred.