It does look like floating point error was it
Particles on the zigs and the zags all look like they are doing
what they're supposed to. The only difference is that for some
particles (a minority) the accumulated current is slightly
different between the condensed branch and move_p_zigzags_v2.
Sometimes the difference (almost always in the last decimal place),
is greater, sometimes less, but it appears to be systematically in
advance_p rather than in move_p. Maybe it has something to do with
the local particle movers?

Moving towards profiling on Darwin.