The periodic boundaries appear to work
Grepping the output from harris for the final cell position and
diffing the results between this branch and devel yields nothing.
So it looks right. The energies are really close as well for a few
time steps, and then it explodes. So that's not great. It explodes
for the diag_flow deck too.

I'm going to make a new zigzag_condensed branch and then merge this
one with that one. Maybe it has to do with an error in the old
zigzag code?