The harris deck seems to agree now
...through 160ish of the 242 steps simulated in serial. From there
it is actually systematically lower than the devel branch, making
me think that there is less systematic heating with zigzags now!
Incidentally in the diag_flow deck, the energies are comparable
but not identical. This discrepancy could be attributed though to
the non-physical properties of the deck. In a physical deck, like
harris, things are pretty close and very exciting.