Started the MOVE_P_STUDY on devel
This study will count the number of times particles will leave the
cell for comparison with zigzags as a function of temperature (i.e.
as a function of average particle velocity). The study will count
the first time a particle leaves the cell, if it does at all, per
time step. It will not count ALL times per time step , as the
zigzag algorithm will always only allow a particle to cross the
boundary one time per time step.