Added a final position test
This greps for the differences in each particle's final cell
number, final position, and final velocity. Here, final means
the values right at the end of the parallel_for in advance_p_
kokkos(...). The test is toggled on with a CMake flag, but is
by default OFF.

The cell number and velocities are identical between zigzags
and devel. This seems to imply all neighbors are appropriately
accounted for and the velocities are unchanged by the zigzags
(as they should be). Many of the final particle positions
differ in the final bit of the mantissa (so there is floating
point error).