The PBCs appear to be working but the neighbors are not
In move_p_kokkos(...) when the particles hit a reflective boundary
along more than one axis, the resulting motion is complicated.
Basically it looks like the best way to do it is by repeating the
zig step until there are no longer and zigs left. There are still
issues with this to be resolved.